+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/DemonHunter/AllFelBreaksLoose.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.DemonHunter;
+
+public class AllFelBreaksLoose: ICardWithRelatedCards
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Demonhunter.AllFelBreaksLoose;
+
+	public bool ShouldShowForOpponent(Player opponent) => false;
+
+	public List<Card?> GetRelatedCards(Player player) =>
+		player.DeadMinionsCards
+			.Select(entity => CardUtils.GetProcessedCardFromEntity(entity, player))
+			.Where(card => card != null && card.IsDemon())
+			.OrderByDescending(card => card!.Cost)
+			.ToList();
+}