+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Hunter/AbominableBowman.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Hunter;
+
+public class AbominableBowman: ICardWithRelatedCards
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Hunter.AbominableBowman;
+
+	public bool ShouldShowForOpponent(Player opponent) => false;
+
+	public List<Card?> GetRelatedCards(Player player) =>
+		player.DeadMinionsCards
+			.Select(entity => CardUtils.GetProcessedCardFromEntity(entity, player))
+			.Where(card => card != null && card.IsBeast())
+			.Distinct()
+			.OrderByDescending(card => card!.Cost)
+			.ToList();
+}