+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Paladin/AnyfinCanHappen.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Paladin;
+
+public class AnyfinCanHappen: ICardWithRelatedCards
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Paladin.AnyfinCanHappen;
+
+	public bool ShouldShowForOpponent(Player opponent) => false;
+
+	public List<Card?> GetRelatedCards(Player player) =>
+		player.DeadMinionsCards
+			.Select(entity => CardUtils.GetProcessedCardFromEntity(entity, player))
+			.Where(card => card != null && card.IsMurloc())
+			.OrderByDescending(card => card!.Cost)
+			.ToList();
+}