+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Priest/Resurrect.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Priest;
+
+public class Resurrect: ICardWithRelatedCards
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Priest.Resurrect;
+
+	public bool ShouldShowForOpponent(Player opponent) => false;
+
+	public List<Card?> GetRelatedCards(Player player) =>
+		player.DeadMinionsCards
+			.Select(entity => CardUtils.GetProcessedCardFromEntity(entity, player))
+			.Where(card => card != null)
+			.Distinct()
+			.OrderByDescending(card => card!.Cost)
+			.ToList();
+}