+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Warlock/BloodreaverGuldan.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Warlock;
+
+public class BloodreaverGuldan: ICardWithRelatedCards
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Warlock.BloodreaverGuldan;
+
+	public bool ShouldShowForOpponent(Player opponent)
+	{
+		var card = Database.GetCardFromId(GetCardId());
+		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentFormat, opponent.OriginalClass) && GetRelatedCards(opponent).Count > 1;
+	}
+
+	public List<Card?> GetRelatedCards(Player player) =>
+		player.DeadMinionsCards
+			.Select(entity => CardUtils.GetProcessedCardFromEntity(entity, player))
+			.Where(card => card is not null && card.IsDemon())
+			.OrderByDescending(card => card!.Cost)
+			.ToList();
+}